minor merge changes
