+++ cpp/skyweaver/SkyCleaver.hpp
@@ -29,18 +29,14 @@ struct BridgeReader {
     std::string freq;
 
 }; // BridgeReader
-
-
+template <typename InputVectorType, typename OutputVectorType>
 class SkyCleaver
 {
   public:
-    using InputType        = int8_t;
-    using OutputType       = uint8_t;
-    using InputVectorType  = TDBPowersStdH<InputType>;
-    using OutputVectorType = TFPowersStdH<OutputType>;
     using FreqType         = std::size_t; // up to the nearest Hz
     using BeamNumberType   = std::size_t;
     using DMNumberType     = std::size_t;
+    using StokesNumberType = std::size_t;
 
   private:
     SkyCleaverConfig const& _config;
@@ -53,17 +49,24 @@ class SkyCleaver
     ObservationHeader _header;
 
     std::vector<std::string> _beam_filenames;
-    std::map<DMNumberType,
-             std::map<BeamNumberType,
-                      std::unique_ptr<MultiFileWriter<OutputVectorType>>>>
+    std::map<
+        StokesNumberType,
+        std::map<DMNumberType,
+                 std::map<BeamNumberType,
+                          std::unique_ptr<MultiFileWriter<OutputVectorType>>>>>
         _beam_writers;
-    std::map<DMNumberType,
-             std::map<BeamNumberType, std::shared_ptr<OutputVectorType>>>
-        _beam_data;
+
+    std::map <StokesNumberType,
+        std::map<DMNumberType,
+                 std::map<BeamNumberType, std::shared_ptr<OutputVectorType>>>>
+            _beam_data;
     std::size_t _total_beam_writers;
 
+    int _nthreads_read;
     void init_readers();
     void init_writers();
+    void read(std::size_t gulp_samples);
+    void write();
 
     Timer _timer;
 
@@ -71,11 +74,12 @@ class SkyCleaver
     SkyCleaver(SkyCleaverConfig const& config);
     SkyCleaver(SkyCleaver const&)     = delete;
     void operator=(SkyCleaver const&) = delete;
-
     void cleave();
 
 }; // class SkyCleaver
 
+
 } // namespace skyweaver
+#include "skyweaver/detail/SkyCleaver.cpp"
 
 #endif // SKYWEAVER_SKYCLEAVER_HPP
